# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WKIL Backend API
  version: 1.0.0
  license:
    name: ISC
    identifier: ISC
  description: Contract for the Express v1 backend, including auth, Meta channels, content, media, widgets, admin, manager, analytics, integrations, and operational endpoints.
servers:
  - url: /
    description: Current API host
security:
  - bearerAuth: []
  - accessTokenCookie: []
...
paths:
...
  /v1/mobile/auth/login:
    post:
      tags: [Auth]
      operationId: mobileLogin
      summary: Log in from a native mobile client
      security: []
      requestBody: { $ref: "#/components/requestBodies/Login" }
      responses:
        "200":
          description: Mobile authentication session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MobileAuthSessionResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /v1/mobile/auth/logout:
    post:
      tags: [Auth]
      operationId: mobileLogout
      summary: Log out a native mobile client
      description: >-
        Supply the refresh token in the Authorization bearer header or JSON
        refreshToken field to revoke the session. Existing refreshToken cookies
        remain supported. A nonblank JSON token takes precedence, then the cookie,
        then the bearer header. Logout remains idempotent and succeeds when no
        credential is supplied or the supplied credential cannot be decoded.
      security:
        - mobileRefreshBearer: []
        - refreshTokenCookie: []
        - {}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string, minLength: 1 }
      responses:
        "200": { $ref: "#/components/responses/MessageOk" }
  /v1/mobile/auth/me:
    get:
      tags: [Auth]
      operationId: mobileCurrentUser
      summary: Get the current native mobile user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current mobile user.
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: "#/components/schemas/MobileUser" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Rotate refresh token and return a new access token
      security:
        - refreshTokenCookie: []
      responses:
        "200": { $ref: "#/components/responses/AuthOk" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    mobileRefreshBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: A WKIL refresh token, not an access token, for mobile refresh and logout.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    accessTokenCookie:
      type: apiKey
      in: cookie
      name: accessToken
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refreshToken
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
    widgetSiteKey:
      type: apiKey
      in: header
      name: X-Widget-Site-Key
...
  responses:
...
    AuthOk:
      description: Auth response.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/AuthResponse" }
...
    MessageOk:
      description: Message response.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/MessageResponse" }
...
    BadRequest:
      description: Bad request or validation failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Authentication required or failed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  requestBodies:
...
    Login:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/LoginRequest" }
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: true
      properties:
        status: { type: string }
        message: { type: string }
        code: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
    MessageResponse:
      type: object
      additionalProperties: true
      properties:
        message: { type: string }
        accessToken: { type: string }
        status: { type: string }
...
    AuthResponse:
      type: object
      additionalProperties: true
      properties:
        user: { $ref: "#/components/schemas/User" }
        message: { type: string }
        accessToken: { type: string }
    MobileUser:
      type: object
      required: [id, email, name, role, avatar, isEmailVerified, isSocialUser]
      properties:
        id: { type: integer }
        email: { type: [string, "null"], format: email }
        name: { type: string }
        role: { type: string }
        avatar: { type: [string, "null"], format: uri }
        isEmailVerified: { type: boolean }
        isSocialUser: { type: boolean }
    MobileAuthSessionResponse:
      type: object
      required: [message, user, accessToken, refreshToken, expiresIn]
      properties:
        message: { type: string }
        user: { $ref: "#/components/schemas/MobileUser" }
        accessToken: { type: string, minLength: 1 }
        refreshToken: { type: string, minLength: 1 }
        expiresIn: { type: integer, minimum: 1 }
...
    User:
      type: object
      additionalProperties: true
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [user, manager, admin, super_admin] }
        avatar: { type: string, format: uri }
        isActive: { type: boolean }
        isEmailVerified: { type: boolean }
        isBusinessProfileCreated: { type: boolean }
        plan: { type: string }
        monthlyCreditsUsed: { type: number }
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }
...